Baud Rate and No Sensor Connect Changes
Change 1.  UART output Baud Rate decreased to 9600 (for standard
usage... so user doesn't need to look up the baud rate of this board).
Change 2. Default LED output for "No device connected" is now "All ON"
because we don't have a power indicator LED.  This will show that the
board is powered when there is no device.  If device is connected, then
the UART output will be working, thus no need for additional change.